refactor(spec): extract normalize-input types from normalize.ts
Move builder-level ChannelInput/AesInput/*LayerInput/SpecInput into
normalize-input.ts so the input surface is separate from canonicalize
logic. Public ./normalize.js re-exports keep index, builder, and tests
on a stable path.
